Fixed (worked around?) transition issue
Transitions were somehow relative to points on the screen that I could
not fully understand. It is related to how the Composition class builds
the composite transitions for the clips.

I added 'align=0' and that seemed to clear the problem up. However, I
do need to go back and research how and why that does what it does.